clean up and reorganize
