+++ src/index.ejs
@@ -20,11 +20,6 @@
     <script async type="text/javascript" id="_fed_an_ua_tag"
         src="https://dap.digitalgov.gov/Universal-Federated-Analytics-Min.js?agency=TREAS&subagency=FS&yt=true&htmlvideo=true"></script>
 
-<script
-  crossOrigin="anonymous"
-  src="//unpkg.com/react-scan/dist/auto.global.js"
-></script>
-
     <!-- Start Verint Survey -->
     <% if (IS_PROD) { %>
         <script type="text/javascript">
