+++ src/js/features/state/overview/StateOverview.jsx
@@ -10,10 +10,10 @@ import {
 } from "data-transparency-ui";
 import { useSelector } from "react-redux";
 
-import GeoVisualizationSectionContainer from './geo/containers/GeoVisualizationSectionContainer';
+import GeoVisualizationSectionContainer from './geo/GeoVisualizationSectionContainer';
 import SummaryStats from './SummaryStats';
-import AwardBreakdownContainer from './awardBreakdown/containers/AwardBreakdownContainer';
-import AwardBreakdownHeader from "./awardBreakdown/AwardBreakdownHeader";
+import AwardBreakdownContainer from './awardBreakdown/AwardBreakdownContainer';
+import AwardBreakdownHeader from "./awardBreakdown/treemap/AwardBreakdownHeader";
 import AwardBreakdownCard from "./awardBreakdown/card/AwardBreakdownCard";
 
 const StateOverview = () => {
